- John Engelhart's NSAssertionHandler implementation, NSObjCRuntime fixes, NSZone NS*Integer fixes, Darwin virtual memory functions - Changed Linux compiler to "default" - Enabled ObjC exceptions in compiler for Windows target

+++ Foundation/NSZone/NSVirtualMemory.m
@@ -1,4 +1,4 @@
-/* Copyright (c) 2006-2007 Christopher J. W. Lloyd
+/* Copyright (c) 2006-2008 Christopher J. W. Lloyd
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
 
@@ -6,14 +6,13 @@ The above copyright notice and this permission notice shall be included in all c
 
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
 
-// Original - Christopher Lloyd <cjwl@objc.net>
 #import <Foundation/NSZone.h>
 
 // NSPageSize implemented in platform subproject
 
 
-unsigned NSLogPageSize(void) {
-   unsigned log,pageSize=NSPageSize();
+NSUInteger NSLogPageSize(void) {
+   NSUInteger log, pageSize=NSPageSize();
 
    for(log=0;(pageSize&1)==0;log++,pageSize>>=1)
     ;
@@ -22,15 +21,15 @@ unsigned NSLogPageSize(void) {
 }
 
 
-unsigned NSRoundDownToMultipleOfPageSize(unsigned byteCount) {
-   unsigned pageSize=NSPageSize();
+NSUInteger NSRoundDownToMultipleOfPageSize(NSUInteger byteCount) {
+   NSUInteger pageSize=NSPageSize();
 
    return (byteCount/pageSize)*pageSize;
 }
 
 
-unsigned NSRoundUpToMultipleOfPageSize(unsigned byteCount) {
-   unsigned pageSize=NSPageSize();
+NSUInteger NSRoundUpToMultipleOfPageSize(NSUInteger byteCount) {
+   NSUInteger pageSize=NSPageSize();
 
    return ((byteCount+(pageSize-1))/pageSize)*pageSize;
 }
